refactor: HttpRequest 생성자 변경 및 패키지 변경
