+++ src/main/java/webserver/RequestHandler.java
@@ -5,7 +5,6 @@ import java.net.Socket;
 import java.nio.file.Files;
 
 import controller.UserController;
-import http.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import util.HttpRequestUtils;
@@ -26,11 +25,8 @@ public class RequestHandler implements Runnable {
                 connection.getPort());
 
         try (InputStream in = connection.getInputStream(); OutputStream out = connection.getOutputStream()) {
-            String firstLine = HttpRequestUtils.getFirstLine(in);
-            String url = HttpRequestUtils.getUrl(firstLine);
-            String method = HttpRequestUtils.getMethod(firstLine);
-            String version = HttpRequestUtils.getVersion(firstLine);
-            HttpRequest httpRequest = new HttpRequest(url, method, version);
+            HttpRequest httpRequest = new HttpRequest(in);
+            String url = httpRequest.getRequestPath();
 
             if (url.startsWith("/user/create")) {
                 url = UserController.createUser(url);
